requirements.txt
imaplib
email
python-dotenv
crewai
